split cache and readers
